+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## 6/30/26
+## TBD
+
+- Adds GA4 `nav_click` event [ISW-5944](https://newyorkpubliclibrary.atlassian.net/browse/ISW-5944)
+
+## v1.4.0 - 6/30/26
 
 - Adds dependabot.yml file for security scanning.
 - Replaces `Connect` link with `Visit` and updates tests.[ISW-5300]